refactor: write.ts 모듈화, axios.ts 수정, detail UI 개선
