+++ api/dbinit/team-brunch/data.json
@@ -203,24 +203,28 @@
       },
       "image": "https://images.pexels.com/photos/68507/spring-flowers-flowers-collage-floral-68507.jpeg?auto=compress&cs=tinysrgb&dpr=1&w=500",
       "content": "LLM(Large Language Model)을 활용하여 작업할 때 토큰 소비 없이 최적의 추론 단계를 유도하는 방식은 매우 중요합니다.\n\n이 기법은 모델에게 복잡한 내부 사고 과정을 직접 출력하도록 요구하지 않으면서도, 정확하고 논리적인 결론 중심의 결과물을 안정적으로 얻게 합니다. 이는 모델의 효율성을 극대화하고, 불필요한 연산 비용(토큰)을 절감하는 핵심 전략입니다.\n\nChain-of-Thought(CoT) 프롬프팅은 모델에게 중간 추론 단계를 명시적으로 출력하도록 요구하여 복잡한 문제 해결 능력을 향상시킵니다. 하지만, 이 중간 단계의 출력 자체가 토큰 소비를 늘려 비용 증가로 이어질 수 있습니다. 특히 대규모 데이터 배치 작업에서는 비용 효율성이 떨어집니다.\n\n이 글에서 소개하는 '사고 과정 프롬프팅 최적화' 기법은, CoT의 이점(높은 정확도와 논리적 추론)을 유지하면서도 출력 토큰을 최소화하는 하이브리드 전략입니다. 즉, 모델에게 내부적으로는 깊은 추론을 하도록 지시하되, 최종 결론만을 간결하게 출력하도록 명령합니다. 이는 'Self-Reflect before Answer' 패턴을 사용하는 것을 포함하며, 출력 토큰을 최소화하는 것이 핵심입니다.\n\n구체적인 방법론으로는 'Self-Reflect before Answer' 패턴을 사용하되, Reflect 단계의 출력 포맷을 JSON이나 XML과 같이 구조화된 형태로 제한하고, 최종 응답 시에는 해당 구조를 삭제하도록 합니다. 이를 통해 모델의 성능을 유지하면서도 출력 비용을 15% 이상 절감한 실제 사례를 분석합니다.\n\n특히 대규모 데이터를 처리하거나, 복잡도가 높은 비즈니스 로직에 적용할 때 탁월한 성능과 일관된 결과 품질을 보장합니다. 간결한 구조 덕분에 유지보수가 용이하고, 다양한 문제 해결에 범용적으로 적용 가능합니다.\n\n결론적으로, 이 사고 과정 프롬프팅 최적화 기법은 작업 효율성을 높이고, 비용을 절감하며, 모델의 잠재력을 최대한 끌어내는 필수적인 AI 활용 전략입니다.\n\n관련하여 심화 과정이나 실제 적용 사례에 대해 더 궁금하시면 언제든 요청해주세요.",
+      "views": 450,
+      "user": {
+        "_id": 4,
+        "name": "프롬프트 마스터",
+        "image": "https://cdn-icons-png.flaticon.com/512/706/706816.png"
+      },
       "replies": [
         {
-          "user": {
-            "_id": 2,
-            "name": "마케터K",
-            "image": "https://cdn-icons-png.flaticon.com/512/6997/6997662.png"
-          },
           "content": "LLM 최적화에 대한 명쾌한 설명 감사합니다.",
-          "like": 15
-        },
-        {
           "user": {
             "_id": 3,
             "name": "철학자J",
             "image": "https://cdn-icons-png.flaticon.com/512/706/706844.png"
-          },
+          }
+        },
+        {
           "content": "기술적인 내용을 쉽게 풀이해주셔서 이해하기 좋았어요!",
-          "like": 15
+          "user": {
+            "_id": 6,
+            "name": "축제매니아",
+            "image": "https://cdn-icons-png.flaticon.com/512/3135/3135789.png"
+          }
         }
       ]
     },
@@ -453,9 +457,9 @@
         },
         {
           "user": {
-            "_id": 4,
-            "name": "프롬프트 마스터",
-            "image": "https://cdn-icons-png.flaticon.com/512/706/706816.png"
+            "_id": 6,
+            "name": "축제매니아",
+            "image": "https://cdn-icons-png.flaticon.com/512/3135/3135789.png"
           },
           "content": "조건을 명확히 하는 과정이 인상적입니다.",
           "like": 15
@@ -525,9 +529,9 @@
         },
         {
           "user": {
-            "_id": 4,
-            "name": "프롬프트 마스터",
-            "image": "https://cdn-icons-png.flaticon.com/512/706/706816.png"
+            "_id": 6,
+            "name": "축제매니아",
+            "image": "https://cdn-icons-png.flaticon.com/512/3135/3135789.png"
           },
           "content": "이름으로 사는 삶, 저도 응원합니다!",
           "like": 15
@@ -561,9 +565,9 @@
         },
         {
           "user": {
-            "_id": 4,
-            "name": "프롬프트 마스터",
-            "image": "https://cdn-icons-png.flaticon.com/512/706/706816.png"
+            "_id": 5,
+            "name": "건강지킴이",
+            "image": "https://cdn-icons-png.flaticon.com/512/706/706831.png"
           },
           "content": "응원합니다! 새로운 보금자리를 찾으시길.",
           "like": 15
@@ -597,9 +601,9 @@
         },
         {
           "user": {
-            "_id": 3,
-            "name": "철학자J",
-            "image": "https://cdn-icons-png.flaticon.com/512/706/706844.png"
+            "_id": 5,
+            "name": "건강지킴이",
+            "image": "https://cdn-icons-png.flaticon.com/512/706/706831.png"
           },
           "content": "글을 통해 위로받는 분들이 많을 거예요.",
           "like": 15
@@ -624,9 +628,9 @@
       "replies": [
         {
           "user": {
-            "_id": 2,
-            "name": "마케터K",
-            "image": "https://cdn-icons-png.flaticon.com/512/6997/6997662.png"
+            "_id": 3,
+            "name": "철학자J",
+            "image": "https://cdn-icons-png.flaticon.com/512/706/706844.png"
           },
           "content": "재미있는 과학 콘텐츠 기대됩니다.",
           "like": 15
@@ -660,9 +664,9 @@
       "replies": [
         {
           "user": {
-            "_id": 2,
-            "name": "마케터K",
-            "image": "https://cdn-icons-png.flaticon.com/512/6997/6997662.png"
+            "_id": 3,
+            "name": "철학자J",
+            "image": "https://cdn-icons-png.flaticon.com/512/706/706844.png"
           },
           "content": "귀촌에 대한 로망이 생기네요.",
           "like": 15
@@ -696,9 +700,9 @@
       "replies": [
         {
           "user": {
-            "_id": 2,
-            "name": "마케터K",
-            "image": "https://cdn-icons-png.flaticon.com/512/6997/6997662.png"
+            "_id": 3,
+            "name": "철학자J",
+            "image": "https://cdn-icons-png.flaticon.com/512/706/706844.png"
           },
           "content": "매주 수요일이 기다려지겠어요!",
           "like": 15
@@ -741,9 +745,9 @@
         },
         {
           "user": {
-            "_id": 3,
-            "name": "철학자J",
-            "image": "https://cdn-icons-png.flaticon.com/512/706/706844.png"
+            "_id": 2,
+            "name": "마케터K",
+            "image": "https://cdn-icons-png.flaticon.com/512/6997/6997662.png"
           },
           "content": "강남 강북 차이가 재밌죠!",
           "like": 15
@@ -777,9 +781,9 @@
         },
         {
           "user": {
-            "_id": 3,
-            "name": "철학자J",
-            "image": "https://cdn-icons-png.flaticon.com/512/706/706844.png"
+            "_id": 2,
+            "name": "마케터K",
+            "image": "https://cdn-icons-png.flaticon.com/512/6997/6997662.png"
           },
           "content": "혼밥의 질을 높이는 꿀팁 부탁드려요.",
           "like": 15
@@ -813,9 +817,9 @@
         },
         {
           "user": {
-            "_id": 3,
-            "name": "철학자J",
-            "image": "https://cdn-icons-png.flaticon.com/512/706/706844.png"
+            "_id": 2,
+            "name": "마케터K",
+            "image": "https://cdn-icons-png.flaticon.com/512/6997/6997662.png"
           },
           "content": "이과 문과 대결, 흥미진진하겠네요.",
           "like": 15
