+++ src/pages/detail-page/detail.css
@@ -21,7 +21,7 @@
   background-position: center;
   filter: blur(8px);
   transform: scale(1.1);
-  opacity: 0.8;
+  opacity: 0.9;
   z-index: -1;
 }
 
@@ -34,13 +34,14 @@
 .post-header > .post-title {
   font-size: 28px;
   line-height: 34px;
-  font-weight: 200;
+  font-weight: 300;
   top: 80px;
 }
 .post-header > .post-subtitle {
   top: 150px;
-  color: #666666;
+  color: rgb(228, 228, 228);
   font-size: 16px;
+  font-weight: 100;
 }
 .post-header > .post-author {
   top: 200px;
@@ -52,7 +53,7 @@
   display: none;
   border: none;
   background: none;
-  color: #888;
+  color: #181818;
   font-size: 14px;
   cursor: pointer;
 }
