+++ src/pages/write-page/modules/authAndNavigation.ts
@@ -1,0 +1,15 @@
+import { getAuthorizationHeader } from '../../../utils/axios';
+
+export function checkAuthAndBackButton() {
+  const auth = getAuthorizationHeader();
+
+  if (!auth || !auth.startsWith('Bearer ')) {
+    window.location.href = '/src/pages/login-page/login.html';
+    return;
+  }
+
+  const backBtn = document.querySelector('.back-btn') as HTMLAnchorElement;
+  backBtn?.addEventListener('click', () => {
+    history.back();
+  });
+}
